# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Arcform AI training API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AuthProfileInputRole = typeof AuthProfileInputRole[keyof typeof AuthProfileInputRole];


export const AuthProfileInputRole = {
  athlete: 'athlete',
  trainer: 'trainer',
} as const;

export interface AuthProfileInput {
  /** @minLength 1 */
  name: string;
  email: string;
  role: AuthProfileInputRole;
}

export type AuthProfileRole = typeof AuthProfileRole[keyof typeof AuthProfileRole];


export const AuthProfileRole = {
  athlete: 'athlete',
  trainer: 'trainer',
} as const;

export interface AuthProfile {
  clerkUserId: string;
  name: string;
  email: string;
  role: AuthProfileRole;
}

export type TrainingProfileInputGoal = typeof TrainingProfileInputGoal[keyof typeof TrainingProfileInputGoal];


export const TrainingProfileInputGoal = {
  strength: 'strength',
  hypertrophy: 'hypertrophy',
  general: 'general',
  fat_loss: 'fat_loss',
} as const;

export type TrainingProfileInputLifestyle = typeof TrainingProfileInputLifestyle[keyof typeof TrainingProfileInputLifestyle];


export const TrainingProfileInputLifestyle = {
  sedentary: 'sedentary',
  active: 'active',
  very_active: 'very_active',
} as const;

export type TrainingProfileInputExperience = typeof TrainingProfileInputExperience[keyof typeof TrainingProfileInputExperience];


export const TrainingProfileInputExperience = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export type TrainingProfileInputEquipment = typeof TrainingProfileInputEquipment[keyof typeof TrainingProfileInputEquipment];


export const TrainingProfileInputEquipment = {
  full_gym: 'full_gym',
  home_gym: 'home_gym',
  bodyweight: 'bodyweight',
} as const;

export type TrainingProfileInputWeightUnit = typeof TrainingProfileInputWeightUnit[keyof typeof TrainingProfileInputWeightUnit];


export const TrainingProfileInputWeightUnit = {
  lb: 'lb',
  kg: 'kg',
} as const;

export type TrainingProfileInputHeightUnit = typeof TrainingProfileInputHeightUnit[keyof typeof TrainingProfileInputHeightUnit];


export const TrainingProfileInputHeightUnit = {
  in: 'in',
  cm: 'cm',
} as const;

export type TrainingProfileInputSex = typeof TrainingProfileInputSex[keyof typeof TrainingProfileInputSex];


export const TrainingProfileInputSex = {
  male: 'male',
  female: 'female',
} as const;

export interface TrainingProfileInput {
  goal: TrainingProfileInputGoal;
  lifestyle: TrainingProfileInputLifestyle;
  experience: TrainingProfileInputExperience;
  /**
     * @minimum 2
     * @maximum 6
     */
  daysPerWeek: number;
  equipment: TrainingProfileInputEquipment;
  /**
     * @minimum 13
     * @maximum 100
     */
  age?: number;
  /** @exclusiveMinimum 0 */
  weight?: number;
  weightUnit?: TrainingProfileInputWeightUnit;
  /** @exclusiveMinimum 0 */
  height?: number;
  heightUnit?: TrainingProfileInputHeightUnit;
  sex?: TrainingProfileInputSex;
  /** @minimum 1 */
  bmrMifflin?: number;
  /** @minimum 1 */
  bmrHarris?: number;
}

export type TrainingProfile = TrainingProfileInput & {
  complete: boolean;
};

export interface GeneratePlanInput {
  profile: TrainingProfileInput;
}

export interface Exercise {
  /**
     * Unique across the workout plan
     * @minLength 1
     */
  id: string;
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  sets: number;
  /** @minLength 1 */
  reps: string;
  /** @minLength 1 */
  targetWeight: string;
  /**
     * @minimum 1
     * @maximum 10
     */
  targetRpe: number;
}

export type WorkoutStatus = typeof WorkoutStatus[keyof typeof WorkoutStatus];


export const WorkoutStatus = {
  upcoming: 'upcoming',
  completed: 'completed',
} as const;

export interface Workout {
  /**
     * Unique across the workout plan
     * @minLength 1
     */
  id: string;
  /** @minLength 1 */
  day: string;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  focus: string;
  /** @minimum 1 */
  duration: number;
  status: WorkoutStatus;
  exercises: Exercise[];
}

export interface TrainingPlan {
  name: string;
  week: number;
  totalWeeks: number;
  rationale: string;
  /** @minItems 1 */
  workouts: Workout[];
}

export interface HistoryItem {
  id: string;
  title: string;
  date: string;
  duration: number;
  totalSets: number;
  avgRpe: number;
}

export interface Dashboard {
  streak: number;
  completedThisWeek: number;
  scheduledThisWeek: number;
  volumeChange: number;
  calorieGoal: number;
  weeklyVolumeGoal: number;
  nextWorkout: Workout;
  history: HistoryItem[];
}

export interface CoachChatInput {
  /** @minLength 1 */
  message: string;
}

export interface CoachReply {
  message: string;
  adjusted: boolean;
}

export interface LoggedSet {
  exerciseId: string;
  setNumber: number;
  reps: number;
  weight: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  rpe: number;
}

export interface WorkoutLogInput {
  sets: LoggedSet[];
  notes?: string;
}

export interface WorkoutLogResult {
  success: boolean;
  adaptation: string;
}

export interface AthleteInviteCode {
  /**
     * @nullable
     * @pattern ^[0-9]{6}$
     */
  code: string | null;
  linked: boolean;
  /** @nullable */
  trainerName?: string | null;
}

export interface NutritionMealInput {
  /** @minLength 1 */
  name: string;
  /** @minimum 0 */
  calories: number;
  /** @minimum 0 */
  protein: number;
  /** @minimum 0 */
  carbs: number;
  /** @minimum 0 */
  fat: number;
}

export type NutritionMeal = NutritionMealInput & {
  id: string;
};

export interface DailyNutrition {
  date: string;
  /** @minimum 0 */
  calorieGoal: number;
  /** @minimum 0 */
  caloriesConsumed: number;
  /** @minimum 0 */
  proteinGoal: number;
  /** @minimum 0 */
  proteinConsumed: number;
  /** @minimum 0 */
  carbsGoal: number;
  /** @minimum 0 */
  carbsConsumed: number;
  /** @minimum 0 */
  fatGoal: number;
  /** @minimum 0 */
  fatConsumed: number;
  meals: NutritionMeal[];
}

export interface TrainerClientLinkInput {
  /** @pattern ^[0-9]{6}$ */
  code: string;
}

export interface TrainerClientSettings {
  /**
     * @minimum 1000
     * @maximum 6000
     */
  calorieGoal?: number;
  /**
     * @minimum 1
     * @maximum 200
     */
  weeklyVolumeGoal?: number;
  /** @minLength 1 */
  workoutSplit?: string;
  /** @minItems 1 */
  workouts?: Workout[];
}

export type TrainerClientUpdateInputType = typeof TrainerClientUpdateInputType[keyof typeof TrainerClientUpdateInputType];


export const TrainerClientUpdateInputType = {
  coach: 'coach',
  progress: 'progress',
  nutrition: 'nutrition',
  programming: 'programming',
} as const;

export interface TrainerClientUpdateInput {
  /** @minLength 1 */
  message: string;
  type: TrainerClientUpdateInputType;
}

export type TrainerClientUpdateType = typeof TrainerClientUpdateType[keyof typeof TrainerClientUpdateType];


export const TrainerClientUpdateType = {
  coach: 'coach',
  progress: 'progress',
  nutrition: 'nutrition',
  programming: 'programming',
} as const;

export interface TrainerClientUpdate {
  id: string;
  date: string;
  message: string;
  type: TrainerClientUpdateType;
}

export type TrainerClientStatus = typeof TrainerClientStatus[keyof typeof TrainerClientStatus];


export const TrainerClientStatus = {
  on_track: 'on_track',
  attention: 'attention',
  new: 'new',
} as const;

export interface TrainerClient {
  id: string;
  name: string;
  email: string;
  goal: string;
  status: TrainerClientStatus;
  lastCheckIn: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  progressScore: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  adherence: number;
  /** @minimum 0 */
  streak: number;
  calorieGoal: number;
  weeklyVolumeGoal: number;
  workoutSplit: string;
  workouts: Workout[];
  updates: TrainerClientUpdate[];
}