# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Arcform AI training API
servers:
  - url: /api
...
paths:
...
  /auth/profile:
    get:
      operationId: getAuthProfile
      tags: [auth]
      responses:
        "200":
          description: Current Arcform profile
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AuthProfile"}
    post:
      operationId: saveAuthProfile
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AuthProfileInput"}
      responses:
        "200":
          description: Saved Arcform profile
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AuthProfile"}
...
components:
  schemas:
...
    AuthProfileInput:
      type: object
      required: [name, email, role]
      properties:
        name: {type: string, minLength: 1}
        email: {type: string, format: email}
        role: {type: string, enum: [athlete, trainer]}
    AuthProfile:
      type: object
      required: [clerkUserId, name, email, role]
      properties:
        clerkUserId: {type: string}
        name: {type: string}
        email: {type: string, format: email}
        role: {type: string, enum: [athlete, trainer]}
...