# lib/api-zod/src/generated/types/authProfileInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Arcform AI training API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthProfileInputRole } from './authProfileInputRole';

export interface AuthProfileInput {
  /** @minLength 1 */
  name: string;
  email: string;
  role: AuthProfileInputRole;
}